+++ src/frontend/src/btc/components/transactions/BtcTransactions.svelte
@@ -13,7 +13,7 @@
 	import { SLIDE_DURATION } from '$lib/constants/transition.constants';
 	import { modalBtcTransaction } from '$lib/derived/modal.derived';
 	import { modalStore } from '$lib/stores/modal.store';
-	import TransactionsSkeleton from '$lib/components/transactions/TransactionsSkeleton.svelte';
+	import TransactionsSkeletons from '$lib/components/transactions/TransactionsSkeletons.svelte';
 
 	let selectedTransaction: BtcTransactionUi | undefined;
 	$: selectedTransaction = $modalBtcTransaction
@@ -23,7 +23,7 @@
 
 <BtcTransactionsHeader />
 
-<TransactionsSkeleton loading={$btcTransactionsNotInitialized}>
+<TransactionsSkeletons loading={$btcTransactionsNotInitialized}>
 	{#each $sortedBtcTransactions as transaction (transaction.data.id)}
 		<div transition:slide={SLIDE_DURATION}>
 			<BtcTransaction transaction={transaction.data} />
@@ -33,7 +33,7 @@
 	{#if $sortedBtcTransactions.length === 0}
 		<TransactionsPlaceholder />
 	{/if}
-</TransactionsSkeleton>
+</TransactionsSkeletons>
 
 {#if $modalBtcTransaction && nonNullish(selectedTransaction)}
 	<BtcTransactionModal transaction={selectedTransaction} />
