+++ src/frontend/src/eth/components/transactions/EthTransactionsSkeletons.svelte
@@ -2,13 +2,13 @@
 	import { erc20UserTokensNotInitialized } from '$eth/derived/erc20.derived';
 	import { ethTransactionsNotInitialized } from '$eth/derived/eth-transactions.derived';
 	import { tokenNotInitialized } from '$eth/derived/nav.derived';
-	import TransactionsSkeleton from '$lib/components/transactions/TransactionsSkeleton.svelte';
+	import TransactionsSkeletons from '$lib/components/transactions/TransactionsSkeletons.svelte';
 
 	let loading: boolean;
 	$: loading =
 		($erc20UserTokensNotInitialized && $tokenNotInitialized) || $ethTransactionsNotInitialized;
 </script>
 
-<TransactionsSkeleton {loading}>
+<TransactionsSkeletons {loading}>
 	<slot />
-</TransactionsSkeleton>
+</TransactionsSkeletons>
