+++ src/frontend/src/icp/components/transactions/IcTransactionsSkeletons.svelte
@@ -1,13 +1,13 @@
 <script lang="ts">
 	import { isNullish } from '@dfinity/utils';
 	import { icTransactionsStore } from '$icp/stores/ic-transactions.store';
-	import TransactionsSkeleton from '$lib/components/transactions/TransactionsSkeleton.svelte';
+	import TransactionsSkeletons from '$lib/components/transactions/TransactionsSkeletons.svelte';
 	import { token } from '$lib/stores/token.store';
 
 	let loading: boolean;
 	$: loading = isNullish($token) || isNullish($icTransactionsStore?.[$token.id]);
 </script>
 
-<TransactionsSkeleton {loading}>
+<TransactionsSkeletons {loading}>
 	<slot />
-</TransactionsSkeleton>
+</TransactionsSkeletons>
